Backend: Implement bid and play logic
Fairly rough implementation of Bid and Play. Player should be able to
place a bid and once all players do so the game moves on to the 'play'
state.
Once the game reaches the first player without any cards, we assume that
all players have no more cards (should always be the case but we should
check later) and the state moves to the 'round end' phase, where it must
calculate the winners and losers.