Frontend: Add StateView class, Add toString methods for Suit and Card
The StateView class contains public aspects of the state needed by the
frontend.

The project method now distinguishes views for the current player and
those for all other players, to make the frontend code easier.

The Suit and Card case classes now have correct toString methods. See
the comment in the Card class to see how the values are treated.